add kwalled (kde) backend and use it if available
the kwallet (kde) backend communicates through dbus with the kwalletd.
it is modeled after the qtkeychain implementation [1].
like qtkeychain, the keyring backend tries to get the kwallet's
'networkWallet' if this fails the 'secret service' is used as before.
kwallet is tried first as the 'secrets service' may exists on systems
running kde but it's very unlikely kwallet is running on non kde
systems.

[1] https://github.com/frankosterfeld/qtkeychain

Signed-off-by: Martin Gysel <me@bearsh.org>