Constraint freebsd and dragonfly to cgo
Both `freebsd` and `drangofly` depend on cgo (via godbus).
This leads to a build failure if you try to build for either of those
targets with CGO_ENABLED=0.

In order to continue to support projects building with `CGO_ENABLED=0`
we add a constraint to `keyring_unix.go` such that it will only be
included for `freebsd` and `dragonfly` when `cgo` is enabled. This means
builds with `CGO_ENABLED=0` will work, but the go-keyring code will not
be included.

Signed-off-by: Mikkel Oscar Lyderik Larsen <mikkel.larsen@zalando.de>