Fix race condition in prompt handling
The current approach to handling prompts involves calling for a prompt
and then registering a signal handler for a response that indicates the
prompt had completed. In most cases, if the user were to take action,
this would be quite slow. However, in the case of automatic responses to
the prompt, the signal response may appear before the signal handler is
configured, resulting in blocking forever.

This commit moves the call for a prompt after the signal handling has
been configured, closing the race window.

Signed-off-by: William Martin <williammartin@github.com>